+++ src/components/atoms/FAQItem.svelte
