+++ src/components/features/Header/ConnectionStatus.svelte
