+++ src/components/features/ProfileOverview/ProfileOverview.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import Profile from '$components/features/ProfileOverview/Profile.svelte';
+  import ActivityLogPreview from '$components/features/ProfileOverview/ActivityLogPreview.svelte';
+</script>
+
+<div class="content-block flex flex-col justify-between gap-2 break-all md:gap-16 lg:flex-row">
+  <Profile />
+  <ActivityLogPreview />
+</div>
