+++ src/components/features/SelectNetworkAndAccount/SelectNetworkAndAccount.svelte
@@ -6,7 +6,7 @@
   import { Account, fetchAccountsForNetwork, type Accounts } from '$lib/stores/accountsStore';
   import type { ApiPromise } from '@polkadot/api';
   import type { web3Enable, web3Accounts } from '@polkadot/extension-dapp';
-  import DropDownMenu from '$components/DropDownMenu.svelte';
+  import DropDownMenu from '$components/atoms/DropDownMenu.svelte';
   import { formatNetwork, formatAccount, isValidURL } from '$lib/utils';
   import { createApi } from '$lib/polkadotApi';
