+++ src/components/pages/Dashboard.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
   import { run } from 'svelte/legacy';
 
-  import Capacity from '$components/Capacity.svelte';
-  import Provider from '$components/Provider.svelte';
-  import DashboardHeader from '$components/DashboardHeader.svelte';
-  import ChainStatus from '$components/ChainStatus.svelte';
+  import Capacity from '$components/features/Capacity/Capacity.svelte';
+  import Provider from '$components/features/Provider/Provider.svelte';
+  import ProfileOverview from '$components/features/ProfileOverview/ProfileOverview.svelte';
   import { fetchAccountsForNetwork } from '$lib/stores/accountsStore';
   import { user } from '$lib/stores/userStore';
   import type { ApiPromise } from '@polkadot/api';
   import { dotApi } from '$lib/stores';
   import { onMount } from 'svelte';
   import type { web3Enable, web3Accounts } from '@polkadot/extension-dapp';
+  import ChainStatus from '$components/features/ChainStatus/ChainStatus.svelte';
 
   let thisWeb3Accounts: typeof web3Accounts | undefined = $state();
   let thisWeb3Enable: typeof web3Enable | undefined = $state();
@@ -31,7 +31,7 @@
 </script>
 
 <div class="max-w-dashboard column w-full" id="dashboard">
-  <DashboardHeader />
+  <ProfileOverview />
 
   <ChainStatus />
