+++ src/components/pages/ProviderLogin.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import BlockSection from '$components/BlockSection.svelte';
-  import LoginForm from './LoginForm.svelte';
-  import HowToTransact from '$components/HowToTransact.svelte';
+  import BlockSection from '$components/atoms/BlockSection.svelte';
+  import LoginForm from '../features/LoginForm/LoginForm.svelte';
+  import HowToTransact from '$components/features/HowToTransact/HowToTransact.svelte';
   import { Button } from '@frequency-chain/style-guide';
   import { goto } from '$app/navigation';
 </script>
