+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import Header from '../components/Header.svelte';
-  import Nav from '../components/Nav.svelte';
+  import Header from '../components/features/Header/Header.svelte';
+  import Nav from '../components/features/Nav/Nav.svelte';
   import { logInPromise, dotApi, storeChainInfo } from '$lib/stores';
   import { getToken } from '$lib/polkadotApi';
   import { getBlockNumber, getEpoch } from '$lib/connections';
