+++ src/routes/[[network=networks]]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import Dashboard from '$components/Dashboard.svelte';
-  import ProviderLogin from '$components/ProviderLogin.svelte';
+  import Dashboard from '$components/pages/Dashboard.svelte';
+  import ProviderLogin from '$components/pages/ProviderLogin.svelte';
   import { pageContent, PageContent } from '$lib/stores/pageContentStore';
 </script>
